requirements.txt
pyyaml~=6.0
requests~=2.28.1
beautifulsoup4~=4.11.1
python-telegram-bot~=20.0a4